Create DownloadButton component
